Major Database change.
The database is now completely redesigned. Now you have logical
separation instead of fifteen fields in one table. Removed old
DatabaseController + Database Views and now we have SeparateDatabase,
which for now is just a basic starting point. Bottle model is left just
for testing purposes and lookups. Probably will be dropped in a commit
or two.